Made refresh better, fixes issues running commands in nonexistent directory

+++ bin/get-and-run.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
 cd ..
-sudo rm -rf $(pwd)
-cd ..
-sudo git clone https://github.com/bvesperman/Sector67RaspberryPiAccessControl.git
-cd Sector67RaspberryPiAccessControl
+if [ -e '.git' ]
+	then git pull
+else
+	sudo rm -rf $(pwd)
+	cd ..
+	sudo git clone https://github.com/bvesperman/Sector67RaspberryPiAccessControl.git
+	cd Sector67RaspberryPiAccessControl
+fi
+git checkout pseudo-master
 DIR=$(pwd)
 os='Linux raspberrypi'
 MAIN=$DIR/space_machines/main.py
